backend/requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.2
bidict==0.22.1
bleach==6.0.0
boto3==1.28.7
botocore==1.31.7
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
debugpy==1.6.6
decorator==5.1.1
deep-translator==1.10.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.1
django-browser-reload==1.11.0
django-cors-headers==4.0.0
django-environ==0.10.0
django-storages==1.13.2
django-tailwind==3.6.0
django-theme==0.3
djangorestframework==3.14.0
docopt==0.6.2
docx2txt==0.8
environ==1.0
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.12.0
fire==0.5.0
Flask==2.2.2
Flask-SocketIO==5.3.2
fonttools==4.39.4
fqdn==1.5.1
future==0.18.3
h11==0.14.0
idna==3.4
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
Js2Py==0.74
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.1
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
opencv-python==4.7.0.72
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pdf2docx==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
#pipwin==0.5.2
platformdirs==3.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg==3.1.9
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
pyjsparser==2.7.1
PyMuPDF==1.22.3
PyPDF2==3.0.1
pypng==0.20220715.0
PyPrind==2.11.3
pyrsistent==0.19.3
pySmartDL==1.3.4
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-engineio==4.3.4
python-json-logger==2.0.5
python-socketio==5.7.2
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.0
qrcode==7.4.2
qtconsole==5.4.0
QtPy==2.3.0
regex==2023.5.5
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
s3transfer==0.6.1
Send2Trash==1.8.0
simple-websocket==0.9.0
six==1.16.0
skimpy==0.0.8
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.3.1
stack-data==0.6.2
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.14
virtualenv==20.23.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wsproto==1.2.0
